chore : 파라미터 오류 및 임시코드 작성

+++ Assets/Scripts/Map/Data/EventData/EnemyData.cs
@@ -11,7 +11,7 @@ public class EnemyData : EventData
     [SerializeField]
     ResourceType resourceType;
     [SerializeField]
-    float resoruceValue;
+    int resoruceValue;
     #endregion
 
     #region Getter & Setter
